requirements.txt
faiss-cpu
joblib
numpy
ollama
pandas
pytest
python-dotenv
scikit-learn
sentence-transformers
streamlit
